% Auto-generated: do not edit by hand
\name{qRCode}

\alias{qRCode}

\title{QRCode component}

\description{
A Dash component for playing a variety of URLs, including file paths, YouTube, Facebook, Twitch, SoundCloud, Streamable, Vimeo, Wistia, Mixcloud, and DailyMotion.
}

\usage{
qRCode(id=NULL, result=NULL, scanDelay=NULL, status=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this compnent in Dash callbacks}

\item{result}{Character. Prop that will contain the result}

\item{scanDelay}{Numeric. Delay for the scanning in ms.}

\item{status}{Character. Contains the status of the renderer}
}

\value{named list of JSON elements corresponding to React.js properties and their values}

